split geom functions and move to shared
As with readimage.c/h there's a lot of useful code that is currently
intertwined with GLUT and GLUI. This splits that code out and put it
into shared.

This is 2/7 patches that are required to parse and post-process .smv
files without GLUT/GLUI etc.
