+++ Source/shared/readgeom.h
@@ -1,0 +1,150 @@
+#ifndef READGEOM_H_DEFINED
+#define READGEOM_H_DEFINED
+#include "MALLOCC.h"
+#include "gd.h"
+#include "options.h"
+#include "string_util.h"
+
+#include <string.h>
+
+#define GEOM_GEOM     0
+#define GEOM_ISO      1
+#define GEOM_SLICE    2
+#define GEOM_BOUNDARY 3
+#define GEOM_CGEOM    4
+
+#define HEADER_SIZE 4
+#define TRAILER_SIZE 4
+#define FORTREAD(var, size, count, STREAM)                                     \
+  FSEEK(STREAM, HEADER_SIZE, SEEK_CUR);                                        \
+  returncode = fread(var, size, count, STREAM);                                \
+  if (returncode != count) returncode = 0;                                     \
+  FSEEK(STREAM, TRAILER_SIZE, SEEK_CUR)
+
+#define FORTREADBR(var, count, STREAM)                                         \
+  FORTREAD(var, 4, (count), STREAM);                                           \
+  if (returncode == 0) break;
+
+/* --------------------------  vertdata ------------------------------------ */
+
+typedef struct _vertdata {
+  float xyz[3], vert_norm[3], texture_xy[3];
+  int itriangle, ntriangles, nused;
+  unsigned char on_mesh_boundary;
+  int geomtype;
+  int isdup;
+  struct _tridata **triangles, *triangle1;
+} vertdata;
+
+/* --------------------------  edgedata ------------------------------------ */
+
+typedef struct _edgedata {
+  int ntriangles;
+  int vert_index[2], itri;
+  vertdata *verts[2];
+  struct _tridata *triangles[2];
+} edgedata;
+
+/* --------------------------  tridata ------------------------------------ */
+
+typedef struct _tridata {
+  unsigned char skinny;
+  float distance, *color, tverts[6], tri_norm[3], vert_norm[9], area;
+  float cface_norm1[3], cface_norm2[3];
+  struct _texturedata *textureinfo;
+  struct _surfdata *geomsurf;
+  struct _geomlistdata *geomlisti;
+  struct _geomobjdata *geomobj;
+  int vert_index[3], exterior, geomtype, insolid, outside_domain;
+  vertdata *verts[3];
+  edgedata *edges[3];
+} tridata;
+
+/* --------------------------  geomlistdata ------------------------------------
+ */
+
+typedef struct _geomlistdata {
+  int nverts, nedges, ntriangles, norms_defined;
+  float *zORIG;
+  vertdata *verts;
+  float *vertvals;
+  tridata *triangles, **triangleptrs, **connected_triangles;
+  edgedata *edges;
+} geomlistdata;
+
+/* --------------------------  geomobjdata ------------------------------------
+ */
+
+typedef struct _geomobjdata {
+  struct _surfdata *surf;
+  struct _texturedata *texture;
+  float *bounding_box, *color;
+  char *texture_name;
+  float texture_width, texture_height, texture_center[3];
+  int texture_mapping;
+  int use_geom_color, ntriangles;
+} geomobjdata;
+
+/* --------------------------  geomdata ------------------------------------ */
+
+typedef struct _geomdata {
+  char *file, *file2, *topo_file;
+  int read_status;
+  int cache_defined;
+  int memory_id, loaded, display;
+  int is_terrain;
+  int block_number;
+  int have_cface_normals, ncface_normals;
+  float *cface_normals;
+  float *float_vals;
+  float bounding_box[6];
+  int *file2_tris, nfile2_tris;
+  int *int_vals, nfloat_vals, nint_vals;
+  float *times;
+  int ntimes, itime, *timeslist;
+  int ngeomobjinfo, geomtype, patchactive, fdsblock;
+  struct _surfdata *surfgeom;
+  geomlistdata *geomlistinfo, *geomlistinfo_0, *currentframe;
+  geomobjdata *geomobjinfo;
+  int *geomobj_offsets;
+  int ngeomobj_offsets;
+} geomdata;
+
+float Distance3(float v1[3], float v2[3]);
+void InitGeom(geomdata *geomi, int geomtype, int fdsblock,
+              int have_cface_normals_arg, int block_number);
+void RotateU2V(float *u, float *v, float *axis, float *angle);
+EXTERNCPP void AngleAxis2Quat(float angle, float *axis, float *quat);
+EXTERNCPP void Quat2Rot(float quat[4], float rot[16]);
+EXTERNCPP void MultQuat(float x[4], float y[4], float z[4]);
+int CompareTransparentTriangles(const void *arg1, const void *arg2);
+EXTERNCPP void XYZ2AzElev(float *xyz, float *azimuth, float *elevation);
+void GetVertType(int nverts, int *triangles, int ntriangles, int *vert_type);
+void AverageVerts2(float v1[3], int v1type, float v2[3], int v2type,
+                   float mesh_bounds[6], float *vavg);
+void AverageVerts3(float v1[3], int v1type, float v2[3], int v2type,
+                   float v3[3], int v3type, float mesh_bounds[6], float *vavg);
+
+#ifdef pp_DECIMATE
+#define IJNODE(i, j) ((j) * nx + (i))
+int PtInTriangle(float *xy, float *v0, float *v1, float *v2, float *zval);
+void DecimateTerrain(vertdata *verts, int nverts, tridata *triangles,
+                     int ntriangles, vertdata **verts_new, int *nverts_new,
+                     tridata **triangles_new, int *ntriangles_new,
+                     float *boxmin, float *boxmax, int nx, int ny);
+void DecimateAllTerrains(void);
+#endif
+
+void GetTriangleNormal(float *v1, float *v2, float *v3, float *normal,
+                       float *area);
+int CompareVerts(const void *arg1, const void *arg2);
+float DistXY(float *x, float *y);
+float GetAngle(float d1, float d2, float d3);
+float GetMinAngle(tridata *trii);
+void InitGeomlist(geomlistdata *geomlisti);
+void ReadGeomFile2(geomdata *geomi);
+void ReadGeomHeader0(geomdata *geomi, int *geom_frame_index, int *ntimes_local);
+void ReadGeomHeader2(geomdata *geomi, int *ntimes_local);
+void ReadGeomHeader(geomdata *geomi, int *geom_frame_index, int *ntimes_local);
+void GetGeomDataHeader(char *file, int *ntimes_local, int *nvals);
+#endif
