+++ Source/smokeview/smokeviewdefs.h
@@ -345,12 +345,6 @@ EXTERNCPP void _Sniff_Errors(const char *whereat, const char *file, int line);
 #define NOT_FDSBLOCK 0
 #define FDSBLOCK     1
 
-#define GEOM_GEOM     0
-#define GEOM_ISO      1
-#define GEOM_SLICE    2
-#define GEOM_BOUNDARY 3
-#define GEOM_CGEOM    4
-
 #define PATCH_STRUCTURED_NODE_CENTER 0
 #define PATCH_STRUCTURED_CELL_CENTER 1
 #define PATCH_GEOMETRY_BOUNDARY      2
